breaking out build tasks into a separate config
